add "total_duration" and "average_duration" as built-in variables of WorldSkill.
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@3575 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.core/src/msi/gama/kernel/simulation/SimulationClock.java
@@ -196,7 +196,7 @@ public class SimulationClock {
 	 * Gets the total duration in milliseconds since the beginning of the current simulation.
 	 * @return a duration in milliseconds
 	 */
-	public static double getTotalDuration() {
+	public static long getTotalDuration() {
 		return total_duration;
 	}
